# src/orders/dto/bulk-delete-order.dto.ts
// src/orders/dto/bulk-delete-order.dto.ts
import { IsArray, ArrayNotEmpty, IsNumber } from 'class-validator';
import { Type } from 'class-transformer';
import { ApiProperty } from '@nestjs/swagger';

export class BulkDeleteOrderDto {
  @ApiProperty({
    description: 'Array of order IDs to delete',
    example: [1, 2, 3],
  })
  @IsArray()
  @ArrayNotEmpty()
  @IsNumber({}, { each: true })
  @Type(() => Number)
  ids: number[];
}