# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/{lang}/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
    "/{lang}/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product or Category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
    "/{lang}/products/search": {
      "get": {
        "operationId": "ProductsController_searchAndFilter",
        "parameters": [
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order (asc or desc)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by (e.g., price, name)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "description": "In-stock status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Category ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No products found"
          }
        },
        "tags": [
          "products"
        ]
      }
    },
...
    "/{lang}/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "accepted",
                "preparing",
                "out_for_delivery",
                "delivered",
                "cancelled"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/{lang}/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "put": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/{lang}/orders/bulk/update": {
      "put": {
        "operationId": "OrdersController_updateMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Bulk update payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManyOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orders updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/{lang}/orders/bulk": {
      "delete": {
        "operationId": "OrdersController_bulkDelete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Array of order IDs to delete",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orders deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Healthy Milk Products API",
    "description": "API for managing milk product orders",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "Product": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Product ID",
            "example": 1
          },
          "quantity": {
            "type": "number",
            "description": "Quantity ordered",
            "example": 2
          },
          "price": {
            "type": "number",
            "description": "Price per unit",
            "example": 6.5
          },
          "size": {
            "type": "string",
            "description": "Optional size/variation",
            "example": "1 liter"
          }
        },
        "required": [
          "id",
          "quantity",
          "price"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "city": {
            "type": "number",
            "description": "City ID for the order",
            "example": 1
          },
          "area": {
            "type": "number",
            "description": "Area ID for the order",
            "example": 2
          },
          "userId": {
            "type": "number",
            "description": "User ID who places the order",
            "example": 1
          },
          "products": {
            "description": "List of products in the order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          },
          "totalAmount": {
            "type": "number",
            "description": "Total order amount (should equal the sum of (price * quantity) for each product)",
            "example": 13
          },
          "paymentMethod": {
            "type": "string",
            "description": "Payment method used (e.g., online, COD)",
            "example": "online"
          },
          "deliveryTime": {
            "type": "string",
            "description": "Delivery time as a string (ISO date)",
            "example": "2025-03-01T12:00:00Z"
          }
        },
        "required": [
          "city",
          "area",
          "userId",
          "products",
          "totalAmount",
          "paymentMethod",
          "deliveryTime"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "city": {
            "type": "number",
            "description": "City ID for the order",
            "example": 1
          },
          "area": {
            "type": "number",
            "description": "Area ID for the order",
            "example": 2
          },
          "userId": {
            "type": "number",
            "description": "User ID who places the order",
            "example": 1
          },
          "products": {
            "description": "List of products in the order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          },
          "totalAmount": {
            "type": "number",
            "description": "Total order amount (should equal the sum of (price * quantity) for each product)",
            "example": 13
          },
          "paymentMethod": {
            "type": "string",
            "description": "Payment method used (e.g., online, COD)",
            "example": "online"
          },
          "deliveryTime": {
            "type": "string",
            "description": "Delivery time as a string (ISO date)",
            "example": "2025-03-01T12:00:00Z"
          }
        }
      },
      "UpdateManyOrderDto": {
        "type": "object",
        "properties": {
          "updates": {
            "description": "Array of order updates",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "updates"
        ]
      },
      "BulkDeleteOrderDto": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "Array of order IDs to delete",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
...
    }
  }
}